refactor(appstore): split controllers and use proper root
