+++ apps/appstore/lib/Search/AppSearch.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
  */
 namespace OCA\Appstore\Search;
 
+use OCA\Appstore\AppInfo\Application;
 use OCP\IL10N;
 use OCP\INavigationManager;
 use OCP\IUser;
@@ -25,7 +26,7 @@ class AppSearch implements IProvider {
 
 	#[\Override]
 	public function getId(): string {
-		return 'settings_apps';
+		return Application::APP_ID;
 	}
 
 	#[\Override]
@@ -35,7 +36,7 @@ class AppSearch implements IProvider {
 
 	#[\Override]
 	public function getOrder(string $route, array $routeParameters): int {
-		return $route === 'settings.AppSettings.viewApps' ? -50 : 100;
+		return $route === 'appstore.Page.viewApps' ? -50 : 100;
 	}
 
 	#[\Override]
