--- apps/appstore/openapi-administration.json
@@ -1,104 +1,0 @@
-{
-    "openapi": "3.0.3",
-    "info": {
-        "title": "settings-administration",
-        "version": "0.0.1",
-        "description": "Nextcloud settings",
-        "license": {
-            "name": "agpl"
-        }
-    },
-    "components": {
-        "securitySchemes": {
-            "basic_auth": {
-                "type": "http",
-                "scheme": "basic"
-            },
-            "bearer_auth": {
-                "type": "http",
-                "scheme": "bearer"
-            }
-        },
-        "schemas": {}
-    },
-    "paths": {
-        "/index.php/settings/admin/log/download": {
-            "get": {
-                "operationId": "log_settings-download",
-                "summary": "download logfile",
-                "description": "This endpoint requires admin access",
-                "tags": [
-                    "log_settings"
-                ],
-                "security": [
-                    {
-                        "bearer_auth": []
-                    },
-                    {
-                        "basic_auth": []
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "Logfile returned",
-                        "headers": {
-                            "Content-Disposition": {
-                                "schema": {
-                                    "type": "string",
-                                    "enum": [
-                                        "attachment; filename=\"nextcloud.log\""
-                                    ]
-                                }
-                            }
-                        },
-                        "content": {
-                            "application/octet-stream": {
-                                "schema": {
-                                    "type": "string",
-                                    "format": "binary"
-                                }
-                            }
-                        }
-                    },
-                    "401": {
-                        "description": "Current user is not logged in",
-                        "content": {
-                            "application/json": {
-                                "schema": {
-                                    "type": "object",
-                                    "required": [
-                                        "message"
-                                    ],
-                                    "properties": {
-                                        "message": {
-                                            "type": "string"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    },
-                    "403": {
-                        "description": "Logged in account must be an admin",
-                        "content": {
-                            "application/json": {
-                                "schema": {
-                                    "type": "object",
-                                    "required": [
-                                        "message"
-                                    ],
-                                    "properties": {
-                                        "message": {
-                                            "type": "string"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    },
-    "tags": []
-}
