+++ lib/private/NavigationManager.php
@@ -245,8 +245,8 @@ class NavigationManager implements INavigationManager {
 					'type' => 'settings',
 					'id' => 'core_apps',
 					'order' => 5,
-					'href' => $this->urlGenerator->linkToRoute('appstore.AppSettings.viewApps'),
-					'icon' => $this->urlGenerator->imagePath('appstore', 'apps.svg'),
+					'href' => $this->urlGenerator->linkToRoute('appstore.Page.viewApps'),
+					'icon' => $this->urlGenerator->imagePath('appstore', 'app.svg'),
 					'name' => $l->t('Apps'),
 				]);
