+++ tests/lib/NavigationManagerTest.php
@@ -279,7 +279,7 @@ class NavigationManagerTest extends TestCase {
 				'id' => 'core_apps',
 				'order' => 5,
 				'href' => '/apps/test/',
-				'icon' => '/apps/settings/img/apps.svg',
+				'icon' => '/apps/appstore/img/app.svg',
 				'name' => 'Apps',
 				'active' => false,
 				'type' => 'settings',
